# back/swagger-spec.yaml
openapi: 3.0.0
paths:
...
  /store:
    post:
      description: Cria uma nova loja para o produtor autenticado
      operationId: StoreController_create
      parameters: []
      requestBody:
        required: true
        description: Dados da loja a ser criada
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateStoreDto'
            examples:
              storeExample:
                summary: Exemplo de loja
                description: Dados básicos para criação de loja
                value:
                  name: Fazenda Orgânica
                  description: Produtos orgânicos cultivados com amor
                  producerProfileId: 1
              storeWithStockExample:
                summary: Loja com estoque inicial
                description: Criar loja com configuração inicial de estoque
                value:
                  name: Horta Feliz
                  description: Produtos frescos direto da horta
                  producerProfileId: 1
                  stock:
                    name: Estoque principal
                    description: Estoque inicial da Horta Feliz
      responses:
        '201':
          description: Loja criada com sucesso
        '400':
          description: Dados inválidos
        '401':
          description: Usuário não autenticado
        '403':
          description: Usuário não tem permissão para criar lojas
      security:
        - JWT-auth: []
      summary: Criar uma nova loja
      tags: &ref_3
        - Store
    get:
      description: Retorna uma lista com todas as lojas cadastradas
      operationId: StoreController_findAll
      parameters: []
      responses:
        '200':
          description: Lista de lojas retornada com sucesso
      summary: Listar todas as lojas
      tags: *ref_3
  /store/{id}:
    get:
      description: Retorna os dados de uma loja específica
      operationId: StoreController_findOne
      parameters:
        - name: id
          required: true
          in: path
          description: ID da loja
          schema:
            type: number
      responses:
        '200':
          description: Loja encontrada com sucesso
        '404':
          description: Loja não encontrada
      summary: Obter loja por ID
      tags: *ref_3
    patch:
      description: Atualiza os dados de uma loja específica
      operationId: StoreController_update
      parameters:
        - name: id
          required: true
          in: path
          description: ID da loja
          schema:
            type: number
      requestBody:
        required: true
        description: Dados para atualização da loja
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateStoreDto'
            examples:
              updateInfoExample:
                summary: Atualizar informações básicas
                description: Atualizar nome e descrição da loja
                value:
                  name: Novo Nome da Loja
                  description: Nova descrição da loja atualizada
              updateRatingExample:
                summary: Atualizar avaliação
                description: Atualizar apenas a avaliação da loja
                value:
                  rating: 4.5
      responses:
        '200':
          description: Loja atualizada com sucesso
        '400':
          description: Dados inválidos
        '401':
          description: Usuário não autenticado
        '403':
          description: Usuário não tem permissão para atualizar esta loja
        '404':
          description: Loja não encontrada
      security:
        - JWT-auth: []
      summary: Atualizar loja
      tags: *ref_3
    delete:
      description: Remove uma loja específica
      operationId: StoreController_remove
      parameters:
        - name: id
          required: true
          in: path
          description: ID da loja
          schema:
            type: number
      responses:
        '200':
          description: Loja removida com sucesso
        '401':
          description: Usuário não autenticado
        '403':
          description: Usuário não tem permissão para remover esta loja
        '404':
          description: Loja não encontrada
      security:
        - JWT-auth: []
      summary: Remover loja
      tags: *ref_3
...
  /orders:
    post:
      description: Cria um novo pedido para um cliente autenticado
      operationId: OrdersController_create
      parameters: []
      requestBody:
        required: true
        description: Dados do pedido a ser criado
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderDto'
            examples:
              basicOrder:
                summary: Pedido básico
                description: Um exemplo de pedido simples
                value:
                  paymentMethod: PIX
                  notes: Entregar no período da tarde, por favor.
              orderWithoutNotes:
                summary: Pedido sem observações
                value:
                  paymentMethod: CREDIT_CARD
      responses:
        '201':
          description: Pedido criado com sucesso
        '400':
          description: Dados inválidos
        '401':
          description: Usuário não autenticado
        '403':
          description: Usuário não tem permissão para criar pedidos
      security: &ref_12
        - JWT-auth: []
      summary: Criar um novo pedido
      tags: &ref_13
        - Orders
    get:
      description: Retorna todos os pedidos do cliente autenticado
      operationId: OrdersController_findAll
      parameters: []
      responses:
        '200':
          description: Lista de pedidos do cliente
        '401':
          description: Usuário não autenticado
        '403':
          description: Usuário não tem permissão para visualizar pedidos
      security: *ref_12
      summary: Listar pedidos do cliente
      tags: *ref_13
  /orders/store:
    get:
      description: Retorna todos os pedidos da loja do produtor autenticado
      operationId: OrdersController_findByStore
      parameters: []
      responses:
        '200':
          description: Lista de pedidos da loja
        '401':
          description: Usuário não autenticado
        '403':
          description: Usuário não tem permissão para visualizar pedidos da loja
      security: *ref_12
      summary: Listar pedidos da loja
      tags: *ref_13
  /orders/{id}:
    get:
      description: Retorna um pedido específico pelo seu ID
      operationId: OrdersController_findOne
      parameters:
        - name: id
          required: true
          in: path
          description: ID do pedido
          schema:
            type: number
      responses:
        '200':
          description: Pedido encontrado
        '401':
          description: Usuário não autenticado
        '403':
          description: Usuário não tem permissão para visualizar este pedido
        '404':
          description: Pedido não encontrado
      security: *ref_12
      summary: Buscar pedido por ID
      tags: *ref_13
    patch:
      description: Atualiza os dados de um pedido existente
      operationId: OrdersController_update
      parameters:
        - name: id
          required: true
          in: path
          description: ID do pedido
          schema:
            type: number
      requestBody:
        required: true
        description: Dados para atualização do pedido
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateOrderDto'
            examples:
              updateAddress:
                summary: Atualizar endereço de entrega
                description: Alterar o endereço de entrega do pedido
                value:
                  notes: Novo endereço de entrega
              updateNotes:
                summary: Atualizar observações
                description: Apenas atualizar as observações do pedido
                value:
                  notes: Por favor, entregar no período da manhã.
      responses:
        '200':
          description: Pedido atualizado com sucesso
        '400':
          description: Dados inválidos
        '401':
          description: Usuário não autenticado
        '403':
          description: Usuário não tem permissão para atualizar este pedido
        '404':
          description: Pedido não encontrado
      security: *ref_12
      summary: Atualizar pedido
      tags: *ref_13
    delete:
      description: Remove um pedido existente
      operationId: OrdersController_remove
      parameters:
        - name: id
          required: true
          in: path
          description: ID do pedido
          schema:
            type: number
      responses:
        '204':
          description: Pedido removido com sucesso
        '401':
          description: Usuário não autenticado
        '403':
          description: Usuário não tem permissão para remover este pedido
        '404':
          description: Pedido não encontrado
      security: *ref_12
      summary: Remover pedido
      tags: *ref_13
  /orders/{id}/status:
    patch:
      description: Atualiza o status de um pedido existente (apenas para produtores)
      operationId: OrdersController_updateStatus
      parameters:
        - name: id
          required: true
          in: path
          description: ID do pedido
          schema:
            type: number
      requestBody:
        required: true
        description: Novo status do pedido
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateOrderStatusDto'
            examples:
              processing:
                summary: Pedido em processamento
                value:
                  status: PROCESSING
              shipped:
                summary: Pedido enviado
                value:
                  status: SHIPPED
              delivered:
                summary: Pedido entregue
                value:
                  status: DELIVERED
      responses:
        '200':
          description: Status do pedido atualizado com sucesso
        '400':
          description: Status inválido
        '401':
          description: Usuário não autenticado
        '403':
          description: Usuário não tem permissão para atualizar status
        '404':
          description: Pedido não encontrado
      security: *ref_12
      summary: Atualizar status do pedido
      tags: *ref_13
  /orders/{id}/cancel:
    patch:
      description: Cancela um pedido existente
      operationId: OrdersController_cancelOrder
      parameters:
        - name: id
          required: true
          in: path
          description: ID do pedido
          schema:
            type: number
      responses:
        '200':
          description: Pedido cancelado com sucesso
        '400':
          description: Pedido não pode ser cancelado (já enviado ou entregue)
        '401':
          description: Usuário não autenticado
        '403':
          description: Usuário não tem permissão para cancelar este pedido
        '404':
          description: Pedido não encontrado
      security: *ref_12
      summary: Cancelar pedido
      tags: *ref_13
info:
  title: Ze da Horta API
  description: API documentation for the NestJS application
  version: '1.0'
  contact: {}
...
servers: []
components:
  securitySchemes:
    JWT-auth:
      scheme: bearer
      bearerFormat: JWT
      type: http
      name: JWT
      description: Insira seu token JWT
      in: header
  schemas:
...
    CreateStoreDto:
      type: object
      properties: {}
    UpdateStoreDto:
      type: object
      properties: {}
...
    CreateOrderDto:
      type: object
      properties: {}
    UpdateOrderDto:
      type: object
      properties: {}
    UpdateOrderStatusDto:
      type: object
      properties: {}